+++ NEWS.md
@@ -11,6 +11,7 @@ Fixes:
   * Prevent crash in residency() if one group has no valid detections at all.
   * Fix bug in plotRatios() brought up in issue [#77](https://github.com/hugomflavio/actel/issues/77) which would occur when the user specified a section for which not all groups were detected.
   * Prevent crashes in residency() when the data has an unusually high percentage of exact-time detections (e.g. midnight).
+  * Prevent crash if only some tags have multiple sensors and Sensor.unit is only included for those tags.
 
 Changes:
   * Removed deprecated argument `plot.detections.by`.